Parameter Builder._

val fetch : log:Build_log.t -> rootfs:string -> string -> Config.env Lwt.t

fetch ~log ~rootfs base initialises the rootfs directory by fetching and extracting the base image. Returns the image's environment.

parameter log
Used for outputting the progress of the fetch

parameter rootfs
The directory in which to extract the base image