Module Docker_sandbox.Docker_config

Derive Docker command-line client parameters from an OBuilder configuration.

val make : Config.t -> ?config_dir:string -> t -> string list * string list

make obuilder_config ~config_dir sandbox_config returns docker_argv, argv where docker_argv is the list of arguments to give to the Docker command-line client, and argv the command to execute in the container.