Module Obuilder.User_temp

The user template fetcher assumes given some base "image" that there is a corresponding directory /Users/<base> and rsyncs this directory over to the provided rootfs directory.

include S.FETCHER

val fetch : log:Build_log.t -> rootfs:string -> string -> Config.env Lwt.t

fetch ~log ~rootfs base initialises the rootfs directory by fetching and extracting the base image. Returns the image's environment.

parameter log
Used for outputting the progress of the fetch

parameter rootfs
The directory in which to extract the base image