Module Obuilder.Zfs_clone

Fetchers

The ZFS Clone fetcher assumes given some base "image" that there is a corresponding ZFS volume obuilder/base-image/<base> and zfs clones this directory over to the provided rootfs.

include S.FETCHER

val fetch : log:Build_log.t -> rootfs:string -> string -> Config.env Lwt.t

fetch ~log ~rootfs base initialises the rootfs directory by fetching and extracting the base image. Returns the image's environment.

parameter log
Used for outputting the progress of the fetch

parameter rootfs
The directory in which to extract the base image