Module Build.Make

Parameters

Store : Obuilder.S.STORE

Sandbox : Obuilder.S.SANDBOX

Signature

include Obuilder.S.BUILDER with type context := Context.t

type t

type context

val build : t -> context -> Obuilder_spec.stage -> (Obuilder.S.id, [ `Cancelled | `Msg of string ]) Lwt_result.t

val delete : ?⁠log:(Obuilder.S.id -> unit) -> t -> Obuilder.S.id -> unit Lwt.t

delete ?log t id removes id from the store, along with all of its dependencies. This is for testing. Note that is not safe to perform builds while deleting: the delete might fail because an item got a new child during the delete, or we might delete something that the build is using.

parameter log

Called just before deleting each item, so it can be displayed.

val prune : ?⁠log:(Obuilder.S.id -> unit) -> t -> before:Unix.tm -> int -> int Lwt.t

prune t ~before n attempts to remove up to n items from the store, all of which were last used before before. Returns the number of items removed.

parameter log

Called just before deleting each item, so it can be displayed.

val v : store:Store.t -> sandbox:Sandbox.t -> t