Module Db_store.Make

Parameters

Raw : Obuilder.S.STORE

Signature

type t

val build : ?⁠switch:Lwt_switch.t -> t -> ?⁠base:Obuilder.S.id -> id:Obuilder.S.id -> log:Obuilder.S.logger -> (cancelled:unit Lwt.t -> log:Obuilder.Build_log.t -> string -> (unit, [ `Cancelled | `Msg of string ]) Lwt_result.t) -> (Obuilder.S.id, [ `Cancelled | `Msg of string ]) Lwt_result.t

build t ~id ~log fn ensures that id is cached, using fn ~cancelled ~log dir to build it if not. If cancelled resolves, the build should be cancelled. If id is already in the process of being built, this just attaches to the existing build.

parameter switch

Turn this off if you no longer need the result. The build will be cancelled if no-one else is waiting for it.

val delete : ?⁠log:(Obuilder.S.id -> unit) -> t -> Obuilder.S.id -> unit Lwt.t

val prune : ?⁠log:(Obuilder.S.id -> unit) -> t -> before:Unix.tm -> int -> int Lwt.t

val result : t -> Obuilder.S.id -> string option

val cache : user:Obuilder_spec.user -> t -> string -> (string * (unit -> unit Lwt.t)) Lwt.t

val wrap : Raw.t -> t