Module Obuilder.Store_spec

Configuration information to set up a store.

type t = [

| `Btrfs of string

| `Zfs of string

| `Rsync of string * Rsync_store.mode

| `Xfs of string

| `Docker of string

]

val is_absolute : string -> bool

val of_string : string -> ([> `Btrfs of string | `Docker of string | `Rsync of string | `Xfs of string | `Zfs of string ], [> `Msg of string ]) Stdlib.result

val pp : Stdlib.Format.formatter -> [< `Btrfs of string | `Docker of string | `Rsync of string | `Xfs of string | `Zfs of string ] -> unit

type store =

| Store : (module S.STORE with type t = 'a) * 'a -> store

val to_store : [< `Btrfs of string | `Docker of string | `Rsync of string * Rsync_store.mode | `Xfs of string | `Zfs of string ] -> [> `Docker | `Native ] * store Lwt.t

val store_t : [ `Btrfs of string | `Docker of string | `Rsync of string | `Xfs of string | `Zfs of string ] Cmdliner.Arg.conv

val store : ?docs:string -> string list -> [ `Btrfs of string | `Docker of string | `Rsync of string | `Xfs of string | `Zfs of string ] option Cmdliner.Arg.t

val rsync_mode_opt : Rsync_store.mode option Cmdliner.Arg.t

val rsync_mode : Rsync_store.mode option Cmdliner.Term.t

val of_t : [> `Btrfs of 'a | `Docker of 'b | `Rsync of 'c | `Xfs of 'd | `Zfs of 'e ] option -> 'f option -> [> `Btrfs of 'a | `Docker of 'b | `Rsync of 'c * 'f | `Xfs of 'd | `Zfs of 'e ]

Transform a store and rsync-mode into a validated combination.

For example an rsync store must supply an rsync-mode.

val v : [> `Btrfs of string | `Docker of string | `Rsync of string * Rsync_store.mode | `Xfs of string | `Zfs of string ] Cmdliner.Term.t

Parse cli arguments for t

val cmdliner : ([> `Docker | `Native ] * store Lwt.t) Cmdliner.Term.t

Parse cli arguments for t and initialise a store.